some bug fixes

+++ resources/views/contents/actor-detail.blade.php
@@ -13,10 +13,12 @@
         <div class="actor-detail-img-and-personal-info">
             <div style="position: relative">
                 <img src="{{ Storage::url('images/actors/'. $actor->image); }}" alt="">
-                <span style="position: absolute; top: 0; right:0;">
-                    <a href="/actor/edit/{{$actor->id}}">Edit</a>
-                    <a href="/actor/{{$actor->id}}/delete">Remove</a>
-                </span>
+                @if (Auth::check() && Auth::user()->role == 'admin')
+                    <span style="position: absolute; top: 0; right:0;">
+                        <a href="/actor/edit/{{$actor->id}}">Edit</a>
+                        <a href="/actor/{{$actor->id}}/delete">Remove</a>
+                    </span>
+                @endif
             </div>
             <div class="personal-info">
                 <h4>Personal Info</h4>
